requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Werkzeug>=2.2
Jinja2
torch>=1.10.0  # pick the version matching your environment (CPU or CUDA)
facenet-pytorch>=2.5.2
opencv-python
Pillow
scikit-learn
joblib
numpy
python-dotenv
